Revert "[clippy] fix println and format part 3"

+++ src/queue/sqs_queue_manager.rs
@@ -73,9 +73,9 @@ impl SqsQueueManager {
         // If custom prefix is provided, use that to construct queue URL
         if let Some(prefix) = &self.config.queue_url_prefix {
             let queue_url = if self.config.fifo {
-                format!("{prefix}/{queue_name}.fifo")
+                format!("{}/{}.fifo", prefix, queue_name)
             } else {
-                format!("{prefix}/{queue_name}")
+                format!("{}/{}", prefix, queue_name)
             };
 
             // Cache the URL
@@ -87,7 +87,7 @@ impl SqsQueueManager {
 
         // Try to get the queue URL from AWS
         let actual_queue_name = if self.config.fifo && !queue_name.ends_with(".fifo") {
-            format!("{queue_name}.fifo")
+            format!("{}.fifo", queue_name)
         } else {
             queue_name.to_string()
         };
@@ -108,7 +108,8 @@ impl SqsQueueManager {
                     Ok(url.to_string())
                 } else {
                     Err(Error::Queue(format!(
-                        "No URL returned for queue: {queue_name}"
+                        "No URL returned for queue: {}",
+                        queue_name
                     )))
                 }
             }
@@ -119,7 +120,7 @@ impl SqsQueueManager {
                 {
                     self.create_queue(queue_name).await
                 } else {
-                    Err(Error::Queue(format!("Failed to get queue URL: {e}")))
+                    Err(Error::Queue(format!("Failed to get queue URL: {}", e)))
                 }
             }
         }
@@ -127,10 +128,10 @@ impl SqsQueueManager {
 
     /// Create a queue if it doesn't exist
     async fn create_queue(&self, queue_name: &str) -> Result<String> {
-        info!("{}", format!("Creating SQS queue: {queue_name}"));
+        info!("{}", format!("Creating SQS queue: {}", queue_name));
 
         let actual_queue_name = if self.config.fifo && !queue_name.ends_with(".fifo") {
-            format!("{queue_name}.fifo")
+            format!("{}.fifo", queue_name)
         } else {
             queue_name.to_string()
         };
@@ -163,7 +164,9 @@ impl SqsQueueManager {
             .set_attributes(Some(attributes))
             .send()
             .await
-            .map_err(|e| Error::Queue(format!("Failed to create SQS queue {queue_name}: {e}")))?;
+            .map_err(|e| {
+                Error::Queue(format!("Failed to create SQS queue {}: {}", queue_name, e))
+            })?;
 
         if let Some(url) = result.queue_url() {
             // Cache the URL
@@ -173,7 +176,8 @@ impl SqsQueueManager {
             Ok(url.to_string())
         } else {
             Err(Error::Queue(format!(
-                "No URL returned after creating queue: {queue_name}"
+                "No URL returned after creating queue: {}",
+                queue_name
             )))
         }
     }
@@ -196,7 +200,10 @@ impl SqsQueueManager {
         tokio::spawn(async move {
             info!(
                 "{}",
-                format!("Starting SQS worker #{worker_id} for queue: {queue_name}")
+                format!(
+                    "Starting SQS worker #{} for queue: {}",
+                    worker_id, queue_name
+                )
             );
 
             let mut interval = interval(Duration::from_secs(1));
@@ -208,7 +215,10 @@ impl SqsQueueManager {
                 if *shutdown.lock().await {
                     info!(
                         "{}",
-                        format!("SQS worker #{worker_id} for queue {queue_name} shutting down")
+                        format!(
+                            "SQS worker #{} for queue {} shutting down",
+                            worker_id, queue_name
+                        )
                     );
                     break;
                 }
@@ -254,20 +264,22 @@ impl SqsQueueManager {
                                                     // Processing succeeded, delete the message
                                                     if let Some(receipt_handle) =
                                                         message.receipt_handle()
-                                                        && let Err(e) = client
+                                                    {
+                                                        if let Err(e) = client
                                                             .delete_message()
                                                             .queue_url(&queue_url)
                                                             .receipt_handle(receipt_handle)
                                                             .send()
                                                             .await
-                                                    {
-                                                        error!(
-                                                            "{}",
-                                                            format!(
-                                                                "Failed to delete message from SQS queue {}: {}",
-                                                                queue_name, e
-                                                            )
-                                                        );
+                                                        {
+                                                            error!(
+                                                                "{}",
+                                                                format!(
+                                                                    "Failed to delete message from SQS queue {}: {}",
+                                                                    queue_name, e
+                                                                )
+                                                            );
+                                                        }
                                                     }
                                                 }
                                                 Err(e) => {
